backend/requirements.txt
# API Framework
fastapi==0.100.0
uvicorn==0.23.2
pydantic==2.1.1
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.19
asyncpg==0.28.0

# AI/ML
tensorflow==2.13.0
face-recognition==1.3.0
opencv-python==4.8.0.74
pillow==10.0.0
scikit-image==0.21.0
numpy==1.24.3
scikit-learn==1.3.0  # For anomaly detection and ML models

# Blockchain
web3==6.8.0
eth-account==0.9.0
py-solc-x==1.1.1
eth-typing==3.4.0
hexbytes==0.3.1

# Zero-Knowledge Proofs
pyzk==0.6.3  # Python wrapper for ZoKrates (ZK framework)
zokrates-pycrypto==0.2.0  # Additional ZK functionality

# Cryptography
cryptography==41.0.3
pycryptodome==3.18.0

# Geo and IP lookups
geoip2==4.7.0
MaxMindDB-geolite2==2018.703  # IP geolocation database

# Anti-fraud
ua-parser==0.16.1  # User-Agent parsing
browser-cookie3==0.19.1  # Browser fingerprinting
fuzzywuzzy==0.18.0  # String matching for fraud detection
python-Levenshtein==0.21.1  # Fast string similarity
pyhocon==0.3.60  # For config files

# Utils
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.0
httpx==0.24.1
pillow==10.0.0
python-dotenv==1.0.0
loguru==0.7.0
pytesseract==0.3.10
pyzbar==0.1.9
pyotp==2.9.0
qrcode==7.4.2
aiofiles==23.1.0
tenacity==8.2.2
psycopg2-binary==2.9.6
alembic==1.11.1
redis==4.6.0
pyjwt==2.8.0
eth-utils==2.1.1
dynaconf==3.2.2 